Refactor exit node management and enhance peer connection logic
- Removed the ExitNode subcommand from the CLI, streamlining command options.
- Updated PeerTableItem to include exit node information, improving data representation.
- Enhanced the logic for determining exit node status based on active connections and metrics.
- Introduced new methods in Peer and PeerMap for evaluating connection convergence and loss rates.
- Improved the handling of exit node data in the PeerManager, ensuring accurate reporting and management.

+++ easytier-core/src/config/mod.rs
@@ -782,7 +782,7 @@ mod tests {
             routes: None,
             peer_policy: None,
             traffic: Some(pb::TrafficConfig {
-                mtu: Some(toml::gen_default_flags().mtu as i32),
+                mtu: Some(toml::gen_default_flags().mtu),
                 instance_recv_bps_limit: Some(100),
                 foreign_relay_bps_limit: None,
             }),
